Fixup: Move file locs after directory tree change
